@@ -6,12 +6,12 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/Foundato/kelon/pkg/constants/logging"
+	"github.com/unbasical/kelon/pkg/constants/logging"
 
-	"github.com/Foundato/kelon/configs"
-	"github.com/Foundato/kelon/pkg/api"
 	"github.com/gorilla/mux"
 	"github.com/pkg/errors"
+	"github.com/unbasical/kelon/configs"
+	"github.com/unbasical/kelon/pkg/api"
 )
 
 type restProxy struct {
