@@ -5,11 +5,11 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/Foundato/kelon/configs"
-	"github.com/Foundato/kelon/pkg/constants"
-	"github.com/Foundato/kelon/pkg/constants/logging"
-	"github.com/Foundato/kelon/pkg/data"
 	"github.com/pkg/errors"
+	"github.com/unbasical/kelon/configs"
+	"github.com/unbasical/kelon/pkg/constants"
+	"github.com/unbasical/kelon/pkg/constants/logging"
+	"github.com/unbasical/kelon/pkg/data"
 
 	// Import mysql dirver
 	_ "github.com/go-sql-driver/mysql"
