@@ -3,10 +3,11 @@ package telemetry
 import (
 	"net/http"
 
-	"github.com/Foundato/kelon/common"
-	"github.com/Foundato/kelon/pkg/constants/logging"
 	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/collectors"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
+	"github.com/unbasical/kelon/common"
+	"github.com/unbasical/kelon/pkg/constants/logging"
 )
 
 type Prometheus struct {
@@ -56,7 +57,7 @@ func (p *Prometheus) Configure() error {
 	if p.registry == nil {
 		p.registry = prometheus.NewRegistry()
 		// System stats
-		p.registry.MustRegister(version, prometheus.NewGoCollector(), prometheus.NewBuildInfoCollector())
+		p.registry.MustRegister(version, collectors.NewGoCollector(), collectors.NewBuildInfoCollector())
 		// Http
 		p.registry.MustRegister(httpRequestsTotal, inFlightRequests, duration, requestSize)
 
